% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rows.r
\name{combine_rows}
\alias{combine_rows}
\title{Function to combine rows based on shared elements}
\usage{
combine_rows(data, sep = ";")
}
\arguments{
\item{data}{A data frame where rows need to be combined.}

\item{sep}{The separator used to split and combine elements (default is ";").}
}
\value{
A list of combined rows.
}
\description{
This function takes a data frame and combines rows that have shared elements. It iterates through each row and checks for shared elements with subsequent rows. If shared elements are found, the rows are combined.
}
\examples{
data <- data.frame(V1 = c("A;B", "C;D", "B;E", "F"), stringsAsFactors = FALSE)
combined <- combine_rows(data, sep = ";")

}
