% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_block_annotation.r
\name{complex_block_annotation}
\alias{complex_block_annotation}
\title{Annotate complex event blocks}
\usage{
complex_block_annotation(
  block,
  window_size = 10000,
  USCS_gene = "TxDb.Hsapiens.UCSC.hg38.knownGene"
)
}
\arguments{
\item{block}{A data frame or tibble containing complex event blocks. Each row should represent a block with "chr" (chromosome) and "start" (start position) columns.}

\item{window_size}{The window size used for creating genomic regions around blocks (default is 10000).}

\item{USCS_gene}{The name of the UCSC transcript database object (default is "TxDb.Hsapiens.UCSC.hg38.knownGene").}
}
\value{
A character vector containing unique gene symbols associated with overlapping blocks.
}
\description{
This function annotates complex event blocks by identifying genes that overlap with the specified genomic regions. It utilizes the GenomicRanges package to perform overlap detection and retrieves gene symbols for the overlapping genes.
}
