% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor_dataframe.r
\name{coor_dataframe}
\alias{coor_dataframe}
\title{Convert junction bin to a dataframe including the first and second junction positions}
\usage{
coor_dataframe(coordinates, window = 10000)
}
\arguments{
\item{coordinates}{A character vector of junction coordinates in the format "chrN:start-chrM:end".}

\item{window}{The window size used for creating genomic regions around junction blocks (default is 10000).}
}
\value{
A list of dataframes, where each dataframe contains information about the first and second junction positions.
}
\description{
This function takes a character vector of junction coordinates and converts it into a dataframe. For each junction, it calculates the positions of the first and second junction blocks based on a specified window size.
}
\examples{
coords <- c("chr1:1000-chr2:2000", "chr3:3000-chr4:4000")
coor_dataframe(coords, window = 5000)

}
