% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplx_events.r
\name{cplx_events}
\alias{cplx_events}
\title{Extract intricate blocks corresponding to each complex event}
\usage{
cplx_events(df, threshold)
}
\arguments{
\item{df}{A dataframe containing genomic blocks and their interactions.}

\item{threshold}{The minimum threshold for the number of interactions required to initiate a complex event.}
}
\value{
A list containing both redundant and non-redundant complex SVs, where each SV is represented by a list containing "blocks" (sorted blocks) and "contigs" (sorted contigs).
}
\description{
This function takes a dataframe of genomic blocks and extracts intricate blocks corresponding to each complex event. It identifies complex events by iteratively collecting interacting blocks and contigs. Both redundant and non-redundant complex structural variations (SVs) are identified and returned.
}
\examples{
data <- data.frame(
  block = c("A", "B", "C"),
  interactions = c("B,C", "A", "A,B")
)
cplx_events(data, threshold = 2)

}
