% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_get.R
\name{gh_get}
\alias{gh_get}
\title{Retrieve information about a file or folder}
\usage{
gh_get(repo_name, org, path)
}
\arguments{
\item{repo_name}{Name of the repository}

\item{org}{GitHub organization}

\item{path}{Path to the file or folder}
}
\value{
A list
}
\description{
Retrieve information about a file or folder
}
\examples{
\donttest{
gh_get("sara", "sara-org", "README.md")
}
}
