% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_get.R
\name{gh_get_repo}
\alias{gh_get_repo}
\title{Retrieve information about a repository}
\usage{
gh_get_repo(repo_name, org)
}
\arguments{
\item{repo_name}{Name of the repository}

\item{org}{Name of the organization}
}
\value{
Response object
}
\description{
Retrieve information about a repository
}
\examples{
\donttest{
gh_get_repo("sara", "sara-org")
}
}
