% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_circos_plot.r
\name{junction_circos_plot}
\alias{junction_circos_plot}
\title{Make circos plot for block junctions}
\usage{
junction_circos_plot(one_complex_ta_cov, link_cut_off = 4, ...)
}
\arguments{
\item{one_complex_ta_cov}{A dataframe containing coverage information for block junctions within a complex event.}

\item{link_cut_off}{The cutoff value for links to be displayed in the circos plot (default is 4).}

\item{...}{Additional arguments (currently not used).}
}
\value{
A circos plot visualizing block junctions and their relationships.
}
\description{
This function generates a circos plot to visualize block junctions and their relationships within a complex event.
}
