% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooe.plot.r
\name{ooe.plot}
\alias{ooe.plot}
\title{Plotting the visual representation of the order of complex events}
\usage{
ooe.plot(res)
}
\arguments{
\item{res}{A bnlearn object representing the learned structure of complex events.}
}
\description{
This function generates a visual representation of the order of complex events using the bnlearn package.
It creates a plot where nodes represent variables (complex events) and their relationships are depicted through directed edges.
Optional highlighting of specific nodes is supported.
}
