% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.of.events.r
\name{order.of.events}
\alias{order.of.events}
\title{Leveraging bnlearn for inferring the order of complex events}
\usage{
order.of.events(
  cplx_list,
  bpdf,
  covdf,
  cplx.event = 1,
  optimization = "pc.stable"
)
}
\arguments{
\item{cplx_list}{A list containing information about complex events, including blocks and contigs.}

\item{bpdf}{A data frame containing breakpoint information, with columns for contig names, start positions, and end positions.}

\item{covdf}{A data frame containing coverage information for the contigs.}

\item{cplx.event}{The index of the complex event to be analyzed.}

\item{optimization}{The type of optimization to perform ("hc" for hill climbing, "pc.stable" for PC-stable algorithm).}
}
\description{
This function uses the bnlearn package to infer the order of complex events based on provided data.
It constructs a Bayesian network using the given complex events, breakpoints, and coverage data to determine their relationships.
}
