% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_bin_mapID.r
\name{overlap_bin_mapID}
\alias{overlap_bin_mapID}
\title{Bin the mapping mapID contig with a resolution of 10kb}
\usage{
overlap_bin_mapID(genome_bin, xmap, window_size = 10000, ...)
}
\arguments{
\item{genome_bin}{A file path or data frame containing genomic bin information with columns "chr", "start", and "end".}

\item{xmap}{A file path or data frame containing xmap information with columns "mapID", "chr", "start_q", "end_q", "start", "end", and "strand".}

\item{window_size}{The size of the binning window in base pairs.}
}
\description{
This function takes genomic bin and xmap data and performs binning with a specified window size.
}
