% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_range.r
\name{within_5kb}
\alias{within_5kb}
\title{Function to check if two ranges are within 5kb of each other}
\usage{
within_5kb(range1, range2)
}
\arguments{
\item{range1}{Numeric value representing the start or end of the first range.}

\item{range2}{Numeric value representing the start or end of the second range.}
}
\value{
TRUE if the two ranges are within 5kb of each other, FALSE otherwise.
}
\description{
This function checks if two ranges are within a specified distance of each other (5kb by default).
}
